refactor selectedItems to use k,v pair instead of repeatedly storing the same item
instead of just putting the n copies of each item in the list.
each item in the basket is a k,v pair of its id and the number of
that item in the basket.

Later this will be used to fetch the items from the main product list
and populate the shopping cart page.